+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.7.0] - TODO
 ### Changed
 - Extracted email verification as its own recipe
+- ThirdParty queries
 
 ## [1.6.0] - 2021-01-14
 ### Changed